Burglaries in Nottingham

Dataset of burglaries (both residential and non-residential) recorded by police in Nottingham, England, in 2022.

Usage

nottingham_burglary

Format

A data frame of 1,795 rows and 5 columns:

month

Date of the first day of the month of the incident

longitude

Longitude

latitude

Latitude

location

Street name or point-of-interest category

lsoa_code

Government Statistical Service code for the Lower Layer Super Output Area in which the incident occurred

This dataset is also available as a stand-alone file in gzipped CSV format at https://mpjashby.github.io/crimemappingdata/nottingham_burglary.csv.gz.

Licence: Open Government Licence

Source

Home Office